+++ inbas-reporting-tool/src/main/webapp/js/utils/DataFilter.js
@@ -1,5 +1,7 @@
 'use strict';
 
+var Constants = require('../constants/Constants');
+
 /**
  * Filters data according to specified filter(s).
  *
@@ -15,19 +17,29 @@ var DataFilter = {
         }
         return data.filter(function (item) {
             for (var key in filter) {
-                if (filter[key] === 'all') {
+                var i, len;
+                if (!filter.hasOwnProperty(key) || filter[key] === Constants.FILTER_DEFAULT) {
                     continue;
                 }
-                var path = key.split('.');
-                var value = item;
-                for (var i = 0, len = path.length; i < len; i++) {
+                var path = key.split('.'),
+                    value = item;
+                for (i = 0, len = path.length; i < len; i++) {
                     value = value[path[i]];
                 }
-                if ((Array.isArray(value) && value.indexOf(filter[key]) === -1) || (!Array.isArray(value) && value !== filter[key])) {
+                var filterValue = filter[key];
+                if (!Array.isArray(filterValue)) {
+                    return (Array.isArray(value) && value.indexOf(filterValue) !== -1) || (!Array.isArray(value) && value === filterValue);
+
+                } else {
+                    // If the filter itself is an array it suffices when a single value from the filter is present
+                    for (i = 0, len = filterValue.length; i < len; i++) {
+                        if (value === filterValue[i] || Array.isArray(value) && value.indexOf(filterValue[i]) !== -1) {
+                            return true;
+                        }
+                    }
                     return false;
                 }
             }
-            return true;
         });
     },
 
@@ -36,7 +48,7 @@ var DataFilter = {
             return true;
         }
         for (var key in filter) {
-            if (filter[key] !== 'all') {
+            if (filter[key] !== Constants.FILTER_DEFAULT) {
                 return false;
             }
         }