+++ inbas-reporting-tool/ReleaseNotes.md
@@ -4,6 +4,10 @@
 - Added support for active/inactive state of a safety issue,
 - Fixed incorrect logo display in Firefox and missing czech translations.
 
+### 0.7.4 - September 29, 2016
+- Report type filter buttons above the reports table (Feature #173).
+- The filter buttons do not show when there is only one kind of reports present.
+
 ### 0.7.3 - September 27, 2016
 - Fixed issues with report cache (Bug #146).
 - Support for more rows per page in tables (Support #152).